add footprint and initialDestination for cesium component
