+++ .eslintrc.js
@@ -12,6 +12,7 @@ module.exports = {
     }],
     "@typescript-eslint/semi": ["error"],
     'vue/no-unused-vars': 'error',
-    'vue/script-setup-uses-vars': 'error'
+    'vue/script-setup-uses-vars': 'error',
+    'vue/require-default-prop': 0
   }
 };
