+++ package.json
@@ -22,6 +22,7 @@
     "cesium": "^1.91.0",
     "django-s3-file-field": "^0.3.0",
     "vue": "^3.2.25",
+    "vue-docgen-api": "^4.44.23",
     "vue-router": "^4.0.12",
     "vuetify": "npm:@vuetify/nightly@next"
   },
