+++ src/store/rgdEndpoints.ts
@@ -1,0 +1,4 @@
+const search = 'rgd/search';
+const rgdSpatialEntry = `rgd/spatial_entry/${spatialID}`;
+const rgdFootprint = `rgd/spatial_entry/${spatialID}/footprint`;
+const rgdImagery = `/rgd_imagery/raster/${spatialID}`;
