+++ src/store/types.ts
@@ -1,0 +1,5 @@
+export type Degrees = {
+  longitude: number,
+  latitude: number,
+  height: number,
+};
