+++ src/stories/CesiumViewer.stories.ts
@@ -1,0 +1,17 @@
+import CesiumViewer  from '@/components/organisms/CesiumViewer.vue';
+
+
+export default {
+  title: 'Organisms/CesiumViewer',
+  component: CesiumViewer,
+};
+
+const Template = (args: any) => ({
+  components: { CesiumViewer },
+  setup() {
+    return { args };
+  },
+  template: '<CesiumViewer v-bind="args"/>',
+});
+
+export const Primary = Template.bind({});
